refactor policy component
